chore: lib 폴더 정리 - 하위 폴더를 밖으로 이동
