+++ src/api/auth.ts
@@ -1,9 +1,11 @@
-import { removeToken } from '@/lib/actions';
+import { removeToken } from '@/actions/token';
+import { auth } from '@/lib/firebase';
 import { User } from '@/types';
+
 import { GoogleAuthProvider, signInWithPopup, signOut } from 'firebase/auth';
 import toast from 'react-hot-toast';
-import { auth } from '../firebase.config';
-import { fetchAPI } from './api';
+
+import { fetchAPI } from './fetcher';
 
 export const loginWithGoogle = async () => {
   const provider = new GoogleAuthProvider();
@@ -30,6 +32,7 @@ export const loginWithGoogle = async () => {
 export const logoutWithGoogle = async () => {
   try {
     await signOut(auth);
+    // TODO: 서버에서 호출하도록 수정
     await removeToken();
   } catch (error: unknown) {
     toast.error('로그아웃에 실패했습니다.');
