+++ src/app/api/books/external/kakao/search/route.ts
@@ -1,6 +1,6 @@
 import { searchBookWithKakaoOpenAPI } from '@/db/controllers';
 import { handleError, parseQuery } from '@/db/utils';
-import { HttpError } from '@/lib/utils';
+import { HttpError } from '@/utils';
 import { NextRequest, NextResponse } from 'next/server';
 
 export async function GET(req: NextRequest) {
