+++ src/app/book/[id]/page.tsx
@@ -1,8 +1,10 @@
+import { getBook, getBookSentence } from '@/api/book';
 import { MaxWidthWrapper } from '@/components/atoms';
 import { Pagination } from '@/components/molecules';
 import { BookInfoSection, SentenceLikeCardList } from '@/components/organisms';
-import { getBook, getBookSentence } from '@/lib/api';
+
 import { Metadata } from 'next';
+
 import classes from './page.module.scss';
 
 type BookDetailPageProps = {
