+++ src/app/components/Header/HeaderMenu/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useUserQuery } from './useUserQuery';
