+++ src/components/molecules/LikeButton/index.tsx
@@ -1,10 +1,12 @@
 'use client';
 
-import { toggleSentenceLike } from '@/lib/api';
+import { toggleSentenceLike } from '@/api/sentence';
 import { Button } from '@mui/material';
+
 import { useState } from 'react';
 import toast from 'react-hot-toast';
 import { FaHeart, FaRegHeart } from 'react-icons/fa6';
+
 import classes from './LikeButton.module.scss';
 
 type LikeButtonProps = {
