+++ src/db/controllers/users.controller.ts
@@ -1,5 +1,5 @@
-import { HttpError } from '@/lib/utils';
 import { PaginationRequest, User } from '@/types';
+import { HttpError } from '@/utils';
 import connectDB from '../connectDB';
 import firebaseAdmin from '../firebase.config';
 import models from '../models';
