--- src/lib/api/index.ts
@@ -1,4 +1,0 @@
-export * from './sentence';
-export * from './book';
-export * from './auth';
-export * from './user';
