+++ src/lib/firebase/index.ts
@@ -1,4 +1,4 @@
-import { initializeApp, FirebaseOptions } from 'firebase/app';
+import { FirebaseOptions, initializeApp } from 'firebase/app';
 import { getAuth } from 'firebase/auth';
 
 import { getStorage } from 'firebase/storage';
